cards

+++ Backend/proyecto/src/main/java/Proyectof/dtos/UsuarioProyecto.java
@@ -3,37 +3,37 @@ package Proyectof.dtos;
 import Proyectof.ConnectionManager;
 
 public class UsuarioProyecto {
-  private long id;
-  private long id_usuario;
-  private long id_proyecto;
+  private int id;
+  private int id_usuario;
+  private int id_proyecto;
 
-  public UsuarioProyecto(long id, long id_usuario, long id_proyecto) {
+  public UsuarioProyecto(int id, int id_usuario, int id_proyecto) {
     this.id = id;
     this.id_usuario = id_usuario;
     this.id_proyecto = id_proyecto;
   }
 
-  public long getId() {
+  public int getId() {
     return id;
   }
 
-  public void setId(long id) {
+  public void setId(int id) {
     this.id = id;
   }
 
-  public long getId_usuario() {
+  public int getId_usuario() {
     return id_usuario;
   }
 
-  public void setId_usuario(long id_usuario) {
+  public void setId_usuario(int id_usuario) {
     this.id_usuario = id_usuario;
   }
 
-  public long getId_proyecto() {
+  public int getId_proyecto() {
     return id_proyecto;
   }
 
-  public void setId_proyecto(long id_proyecto) {
+  public void setId_proyecto(int id_proyecto) {
     this.id_proyecto = id_proyecto;
   }
 }
